not adding extra item when multiple is added, adding up total still
