+++ rest-api/src/main/java/com/acme/dataobjects/Barcodes.java
@@ -4,6 +4,7 @@ package com.acme.dataobjects;
 public class Barcodes {
     private String barcode;
 
+
     public Barcodes(String barcode){
         this.barcode = barcode;
     }
