+++ rest-api/src/main/java/com/acme/dataobjects/ItemResponse.java
@@ -3,19 +3,19 @@ package com.acme.dataobjects;
 import java.util.ArrayList;
 
 public class ItemResponse {
-    private final ArrayList<ProductForList> items;
+    private final ArrayList<Product> items;
 
 
     private final double total;
 
-    public ItemResponse(ArrayList<ProductForList> listOfItems, double total){
+    public ItemResponse(ArrayList<Product> listOfItems, double total){
         this.items = listOfItems;
         this.total = total;
     }
 
 
 
-        public ArrayList<ProductForList> getItem() {
+        public ArrayList<Product> getItem() {
         return items;
     }
